INT-4392: Polishing
JIRA: https://jira.spring.io/browse/INT-4392

Avoid the need for reflection on the `MicrometerMetricsCaptor`.
 - move facade interfaces to a sub-package.
Remove deprecated `MicrometerMetricsFactory.
