+++ spring-integration-core/src/main/java/org/springframework/integration/channel/AbstractPollableChannel.java
@@ -20,8 +20,8 @@ import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.List;
 
-import org.springframework.integration.support.management.CounterFacade;
 import org.springframework.integration.support.management.PollableChannelManagement;
+import org.springframework.integration.support.management.metrics.CounterFacade;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.PollableChannel;
 import org.springframework.messaging.support.ChannelInterceptor;
