+++ spring-integration-core/src/main/java/org/springframework/integration/channel/NullChannel.java
@@ -28,9 +28,9 @@ import org.springframework.integration.support.management.ConfigurableMetricsAwa
 import org.springframework.integration.support.management.DefaultMessageChannelMetrics;
 import org.springframework.integration.support.management.IntegrationManagedResource;
 import org.springframework.integration.support.management.MessageChannelMetrics;
-import org.springframework.integration.support.management.MetricsCaptor;
 import org.springframework.integration.support.management.Statistics;
-import org.springframework.integration.support.management.TimerFacade;
+import org.springframework.integration.support.management.metrics.MetricsCaptor;
+import org.springframework.integration.support.management.metrics.TimerFacade;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.PollableChannel;
 import org.springframework.util.Assert;
