+++ spring-integration-core/src/main/java/org/springframework/integration/endpoint/AbstractMessageSource.java
@@ -26,10 +26,10 @@ import org.springframework.expression.Expression;
 import org.springframework.integration.core.MessageSource;
 import org.springframework.integration.support.AbstractIntegrationMessageBuilder;
 import org.springframework.integration.support.context.NamedComponent;
-import org.springframework.integration.support.management.CounterFacade;
 import org.springframework.integration.support.management.IntegrationManagedResource;
 import org.springframework.integration.support.management.MessageSourceMetrics;
-import org.springframework.integration.support.management.MetricsCaptor;
+import org.springframework.integration.support.management.metrics.CounterFacade;
+import org.springframework.integration.support.management.metrics.MetricsCaptor;
 import org.springframework.integration.util.AbstractExpressionEvaluator;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessagingException;
