+++ spring-integration-core/src/main/java/org/springframework/integration/support/management/IntegrationManagementConfigurer.java
@@ -16,7 +16,6 @@
 
 package org.springframework.integration.support.management;
 
-import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
@@ -33,6 +32,8 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.integration.core.MessageSource;
 import org.springframework.integration.support.management.IntegrationManagement.ManagementOverrides;
+import org.springframework.integration.support.management.metrics.MetricsCaptor;
+import org.springframework.integration.support.management.micrometer.MicrometerMetricsCaptor;
 import org.springframework.integration.support.utils.PatternMatchUtils;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.MessageHandler;
@@ -215,17 +216,7 @@ public class IntegrationManagementConfigurer implements SmartInitializingSinglet
 		ClassLoader classLoader = IntegrationManagementConfigurer.class.getClassLoader();
 		if (ClassUtils.isPresent("io.micrometer.core.instrument.MeterRegistry",
 				classLoader)) {
-			try {
-				// Use reflection to avoid a package tangle with ...management.micrometer
-				Class<?> captor = ClassUtils.forName(
-						"org.springframework.integration.support.management.micrometer.MicrometerMetricsCaptor",
-						classLoader);
-				Method method = captor.getDeclaredMethod("loadCaptor", ApplicationContext.class);
-				this.metricsCaptor = (MetricsCaptor) method.invoke(null, this.applicationContext);
-			}
-			catch (Exception e) {
-				// no op
-			}
+			this.metricsCaptor = MicrometerMetricsCaptor.loadCaptor(this.applicationContext);
 		}
 		if (this.metricsCaptor != null) {
 			injectCaptor();
