+++ spring-integration-core/src/main/java/org/springframework/integration/support/management/metrics/SampleFacade.java
@@ -14,13 +14,15 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.support.management;
+package org.springframework.integration.support.management.metrics;
 
 /**
  * @author Gary Russell
  * @since 5.0.4
  *
  */
-public interface GaugeFacade {
+public interface SampleFacade {
+
+	void stop(TimerFacade timer);
 
 }
