+++ spring-integration-core/src/main/java/org/springframework/integration/support/management/metrics/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides interfaces related to 5.0 metrics.
+ */
+package org.springframework.integration.support.management.metrics;
